+++ apps/block_scout_web/lib/block_scout_web/web_router.ex
@@ -122,6 +122,8 @@ defmodule BlockScoutWeb.WebRouter do
 
     resources("/recent-transactions", RecentTransactionsController, only: [:index])
 
+    resources("/verified-contracts", VerifiedContractsController, only: [:index])
+
     get("/txs", TransactionController, :index)
 
     resources "/tx", TransactionController, only: [:show] do
@@ -144,6 +146,11 @@ defmodule BlockScoutWeb.WebRouter do
       resources("/token-transfers", TransactionTokenTransferController,
         only: [:index],
         as: :token_transfer
+      )
+
+      resources("/state", TransactionStateController,
+        only: [:index],
+        as: :state
       )
     end
 
@@ -497,6 +504,8 @@ defmodule BlockScoutWeb.WebRouter do
     get("/search-logs", AddressLogsController, :search_logs)
 
     get("/search-results", SearchController, :search_results)
+
+    get("/search-verified-contracts", VerifiedContractsController, :search_verified_contracts)
 
     get("/csv-export", CsvExportController, :index)