+++ apps/explorer/config/dev.exs
@@ -1,7 +1,7 @@
 import Config
 
 # Configure your database
-config :explorer, Explorer.Repo.Local, timeout: :timer.seconds(80)
+config :explorer, Explorer.Repo.Local, timeout: :timer.seconds(80), migration_lock: nil
 
 # Configure API database
 config :explorer, Explorer.Repo.Replica1, timeout: :timer.seconds(80)